♻️ move enums to new packages + add JobCategory enum
