+++ users/src/main/kotlin/org/seasonist/entities/Availability.kt
@@ -5,6 +5,7 @@ import io.quarkus.hibernate.orm.panache.kotlin.PanacheEntityBase
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import org.eclipse.microprofile.graphql.Ignore
+import org.seasonist.entities.enums.JobCategory
 import java.util.*
 
 @Entity(name = "availabilities")
@@ -18,5 +19,5 @@ class Availability : PanacheEntityBase {
 	lateinit var dateStart: Date
 	lateinit var dateEnd: Date
 	var searchArea: String? = null
-	var jobCategory: String? = null
+	var jobCategory: JobCategory? = null
 }
