+++ users/src/main/kotlin/org/seasonist/entities/enums/Group.kt
@@ -1,4 +1,4 @@
-package org.seasonist.entities
+package org.seasonist.entities.enums
 
 enum class Group {
 	EMPLOYER {
