+++ users/src/main/kotlin/org/seasonist/entities/enums/JobCategory.kt
@@ -1,0 +1,17 @@
+package org.seasonist.entities.enums
+
+enum class JobCategory {
+	AGRICULTURE_VITICULTURE_PECHERIE,
+	HOTELLERIE_PLEIN_AIR_CLUB_VACANCES_CAMPING_ANIMATION,
+	HOTELS_CAFES_BARS_RESTAURANTS,
+	EVENEMENTIEL,
+	CASINOS_PARCS_ATTRACTION,
+	ADMINISTRATION_ESPACES_CULTURELS_TOURISME,
+	MONTAGNE_SKI,
+	MER_PLONGEE_SPORTS_LOISIRS_NAUTIQUES,
+	SECURITE_GARDIENNAGE,
+	LOGISTIQUE_TRANSPORT,
+	BABY_SITTING_SERVICES_PERSONNE,
+	COMMERCE_ACHATS_VENTE,
+	SPA_ESTHETIQUE_COIFFURE,
+}
