+++ users/src/main/kotlin/org/seasonist/entities/enums/Nationality.kt
@@ -1,4 +1,4 @@
-package org.seasonist.entities
+package org.seasonist.entities.enums
 
 enum class Nationality {
 	FRANCE {
